Use verovio binary to convert, and fall back to musescore if it fails
Sometimes verovio crashes when trying to read a musicxml file, so use
the external binary instead of the python bindings, this prevents the
crash from taking out the entire conversion process.
We've seen that sometimes musescore successfully reads a file that
verovio can't, so use musescore to read and resave the file, then try
again with the verovio conversion.